requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas
numpy
tensorflow
pillow
db-dtypes
pyarrow
scikit-learn
ipdb
ultralytics
comet_ml
fastapi
streamlit
faiss-cpu
